requirements.txt
# Flask and dependencies
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.36
# psycopg2-binary==2.9.10  # Uncomment for PostgreSQL
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Environment
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0
